Still WIP, 'Receiver.rs' removed.
